Translated arrays to fr-rFR Updated to graddle 4.1.3
