+++ FytHWOneKey/debug/output-metadata.json
@@ -1,0 +1,18 @@
+{
+  "version": 2,
+  "artifactType": {
+    "type": "APK",
+    "kind": "Directory"
+  },
+  "applicationId": "org.hvdw.fythwonekey",
+  "variantName": "processDebugResources",
+  "elements": [
+    {
+      "type": "SINGLE",
+      "filters": [],
+      "versionCode": 14,
+      "versionName": "1.4.0",
+      "outputFile": "FytHWOneKey-debug.apk"
+    }
+  ]
+}
