+++ FytHWOneKey/release/output-metadata.json
@@ -1,0 +1,18 @@
+{
+  "version": 2,
+  "artifactType": {
+    "type": "APK",
+    "kind": "Directory"
+  },
+  "applicationId": "org.hvdw.fythwonekey",
+  "variantName": "processReleaseResources",
+  "elements": [
+    {
+      "type": "SINGLE",
+      "filters": [],
+      "versionCode": 14,
+      "versionName": "1.4.0",
+      "outputFile": "FytHWOneKey-release.apk"
+    }
+  ]
+}
