+++ FytHWOneKey/src/main/AndroidManifest.xml
@@ -1,12 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" xmlns:tools="http://schemas.android.com/tools" package="org.hvdw.fythwonekey">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="org.hvdw.fythwonekey">
 
     <!-- <uses-feature android:name="android.hardware.usb.host" /> -->
 
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.CHANGE_WIFI_STATE" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
-    <permission android:name="android.permission.INTERACT_ACROSS_USERS_FULL"/>
+
+    <permission android:name="android.permission.INTERACT_ACROSS_USERS_FULL" />
     <!-- <uses-permission android:name="android.permission.USB_PERMISSION" /> -->
 
     <application
