+++ FytHWOneKey/src/main/java/org/hvdw/fythwonekey/MySettings.java
