+++ FytHWOneKey/src/main/java/org/hvdw/fythwonekey/SettingsActivity.java
