+++ FytHWOneKey/src/main/java/org/hvdw/fythwonekey/activities/OneKeyVoice.java
