+++ FytHWOneKey/src/main/res/logo.xcf
