+++ FytHWOneKey/src/main/res/values-fr-rFR/strings.xml
