+++ FytHWOneKey/src/main/res/values/strings.xml
